build: restructure directories
